Add concurrency support (processes and threads) to FPE spy
This enables tracing into child processes (fork/exec)
and threads (pthread_create).  Each child gets a separate log file.

CAVEAT: there is a concurrency bug in here that is triggered
when we run out of monitoring contexts.   For now make sure
that MAX_CONTEXTS is set large so this does not happen.