Fix RISC-V assumption that all instructions are 4 bytes wide
RISC-V was nice when they defined their compressed instruction set,
and made it so the lowest 2 bits of every NON-COMPRESSED instruction
are 11. Since RISC-V does not use any other formats on those bits
except for compressed instructions, we can do a quick ternary to
figure out the width of the instruction.