Rename exceptmask to enabled fp traps, to match its name to its purpose
In FPSpy, we make the distinction between an "enable" and a "mask",
because FPSpy works with multiple architectures, which means both end
up getting used. x86 uses a mask for FP exceptions whereas ARM and
RISC-V use an enable.

Mask   => If bit is cleared (=0), then event fires. If bit is set (=1),
          then the event is masked off and is NOT seen.
Enable => If bit is set (=1), then the event fires. If the bit is
          clear (=0), then the event is disabled and is NOT seen.

This variable is intended to track the exceptions which the FPSpy user
wants to be SEEN. So we call it an enable instead.