Update RISC-V's clearing of traps to match RISC-V's split FCSR & FFLAGS
On RISC-V, we cannot rely on any library knowing about our new
fflags_care CSR (which is what we want). This means that FPSpy must
directly manipulate ONLY that CSR with csrr/csrw. However, the fcsr
for the processor is provided by both Linux (as part of ucontext) and
is built up by the C-side handler for PPEs when the fake ucontext is
built. Then the data in the ucontext is the one that is written back
to the fcsrs.

This means that ften_base should be IDENTICAL for both fcsr and
fflags_care on our modified RISC-V architecture.