FPSpy properly disables all FP state when FPSpy wants to do FP math
RAFT-V has 2 CSRs for FP support:
  1. FSCSR: The 5 supported IEEE 754 sticky bits are tracked here.
  2. FFLAGS_CARE (CSR 0x880): The 5 enable bits to determine if/which
     of the 5 sticky bits raise precise exceptions.

Sometimes FPSpy wants to do FP math. FPSpy uses it in the sampler to
choose a random time amount of time to sleep from an expoential
distribution, for instance. To make sure FPSpy does not except on
itself, we must save & disable both "pending" FP exceptions AND
disable the raising of FP exceptions when an exception-enabled sticky
bit goes from 0 to 1.